requirements.txt
langchain>=0.2
langchain-core>=0.2
langchain-community>=0.2
langchain-google-genai>=2.1.9

google-generativeai>=0.7.2
google-ai-generativelanguage>=0.6.18,<0.7.0

sentence-transformers>=3.0
torch>=2.0

faiss-cpu>=1.8.0
rank-bm25>=0.2.2

tqdm>=4.66

readline; sys_platform != "win32"

# Optional: re-ranker
# already pulled with sentence-transformers
